Move publish options to dedicated file to be on par with consume_options
